data_proc/pyproject.toml
[project]
name = "data-proc"
version = "0.1.0"
description = "Data processing scripts for BB Kitchen"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "google-generativeai",
    "pillow",
]